Setup versioning (#20)
* Cleaned up the local methods a bit

Still need to add a local get/put object

* Added test@local method

* Some cleanup while retesting the versioning branch

Mostly just improvements to verbose and adding dryrun method

* Some dev improvements

* Deployment improvements

* Fixing README

* Added bom check and some bug fixes

* Added senstate to versioning branch

+++ fetcher/lib/utils.js
@@ -2,12 +2,15 @@ const zlib = require('zlib');
 const { promisify } = require('util');
 const request = promisify(require('request'));
 const AWS = require('aws-sdk');
-const zlib = require('zlib');
 const gzip = promisify(zlib.gzip);
 const unzip = promisify(zlib.unzip);
 const path = require('path');
 const fs = require('fs');
 const csv = require('csv-parser');
+//const stripBomStream = import('strip-bom-stream');
+const readdir = promisify(fs.readdir);
+const mv = promisify(fs.rename);
+
 
 const s3 = new AWS.S3({
   maxRetries: 10
@@ -16,6 +19,8 @@ const s3 = new AWS.S3({
 const { Storage } = require('@google-cloud/storage');
 
 const VERBOSE = !!process.env.VERBOSE;
+const DRYRUN = !!process.env.DRYRUN;
+
 // global storage object
 var storage;
 
@@ -25,6 +30,8 @@ const revokeCredentials = () => {
 
 const applyCredentials = credentials => {
   if(!storage && credentials) {
+    if (!credentials.client_email) throw new Error('client_email required');
+    if (!credentials.client_id) throw new Error('client_id required');
     console.debug(`Initializing storage object '${credentials.project_id}' as ${credentials.client_email}`);
     let projectId = credentials.project_id;
     // https://github.com/googleapis/google-cloud-node/blob/main/docs/authentication.md
@@ -42,19 +49,28 @@ const applyCredentials = credentials => {
  * @returns {object}
  */
 async function fetchSecret(source_name) {
+
+  const region = process.env.AWS_DEFAULT_REGION || 'us-west-2';
+  const stack = process.env.SECRET_STACK || process.env.STACK;
+  const SecretId = `${stack}/${source_name}`;
+  // make sure we have a stack name
+  if (!stack) throw new Error('STACK Env Var Required');
+
   const secretsManager = new AWS.SecretsManager({
-    region: process.env.AWS_DEFAULT_REGION || 'us-east-1'
+    region,
   });
 
-  if (!process.env.STACK) throw new Error('STACK Env Var Required');
-
-  const SecretId = `${process.env.SECRET_STACK || process.env.STACK}/${source_name}`;
-
-  if (VERBOSE) console.debug(`Fetching ${SecretId}...`);
+  if (VERBOSE) console.debug(`Fetching secret - ${region}/${SecretId}...`);
 
   const { SecretString } = await secretsManager.getSecretValue({
     SecretId
-  }).promise();
+  }).promise()
+        .catch( err => {
+          // this is a stop gap until there is a way to say there is
+          // nothing to lookup
+          if (VERBOSE) console.debug(err);
+          return { SecretString: '{}' };
+        });
 
   return JSON.parse(SecretString);
 }
@@ -74,12 +90,13 @@ const getObject = async (Key) => {
 
 const putObject = async (data, Key) => {
   const Bucket = process.env.BUCKET;
-  //console.debug('PUTTING OBJECT',  data.constructor.name, `${Bucket}/${Key}`);
+  var ContentType = 'application/json';
 
   if(data.constructor.name === 'Object') {
     data = JSON.stringify(data);
   } else if(data.constructor.name === 'Measures') {
     data = await gzip(data.csv());
+    ContentType = 'text/csv';
   } else {
     console.warn('Writing object of unknown type',  data.constructor.name, typeof data, typeof data === 'string', data);
   }
@@ -88,22 +105,40 @@ const putObject = async (data, Key) => {
     data = await gzip(data);
   }
 
-  await s3.putObject({
-    Bucket,
-    Key,
-    Body: data,
-    ContentType: 'application/json',
-    ContentEncoding: 'gzip',
-  }).promise();
+  if(!DRYRUN) {
+    if (VERBOSE) console.debug(`Saving data to ${Key}`);
+    await s3.putObject({
+      Bucket,
+      Key,
+      Body: data,
+      ContentType,
+      ContentEncoding: 'gzip',
+    }).promise().catch( err => {
+      console.log('error putting object', err);
+    });
+  } else {
+    if (VERBOSE) console.debug(`Would have saved data to ${Key}`);
+    await writeJson(data, Key);
+  }
 };
 
 const writeJson = async (data, filepath) => {
-  console.debug('writing data to text', filepath);
-  const jsonString = path.extname(filepath) === ".gz"
-        ? await gzip(JSON.stringify(data))
-        : JSON.stringify(data);
+  const dir = process.env.LOCAL_DESTINATION_BUCKET || __dirname;
+  if (VERBOSE) console.debug('writing data to local directory', dir, filepath);
+  let jsonString = data;
+  if(typeof(data) == 'object' && !Buffer.isBuffer(data)) {
+    jsonString = path.extname(filepath) === ".gz"
+      ? await gzip(JSON.stringify(data))
+      : JSON.stringify(data);
+  }
   //await fs.promises.writeFile(filepath, jsonString);
-  fs.writeFileSync(filepath, jsonString);
+  const fullpath = path.join(dir, filepath);
+  fs.promises.mkdir(path.dirname(fullpath), {recursive: true})
+    .then( res => {
+      fs.writeFileSync(fullpath, jsonString);
+    }).catch( err => {
+      console.warn(err);
+    });
   return true;
 };
 
@@ -149,28 +184,35 @@ const listFilesGoogleBucket = async config => {
   const f = [];
 
   const options = {
-	  prefix: 'pending/',
+	  prefix: config.prefix || 'pending/',
 	  delimeter: '/',
   };
 
-  console.debug(`Fetching file list from '${bucket}'`);
+  if (VERBOSE) console.debug(`Fetching file list from '${bucket}/${options.prefix}'`);
   const [files] = await storage.bucket(bucket).getFiles(options);
 
   files.forEach(file => {
-	  f.push({
-      source: 'google-bucket',
-	    path: file.name,
-	    name: path.basename(file.name),
-	    id: file.id,
-	    bucket,
-	  });
+    if(!file.name.endsWith('/')) {
+      if (VERBOSE) console.debug('listing file', file.name);
+	    f.push({
+        source: 'google-bucket',
+	      path: file.name,
+	      name: path.basename(file.name),
+	      id: file.id,
+	      bucket,
+	    });
+    } else {
+      if (VERBOSE) console.debug('Skipping directory path', file.name);
+    }
   });
   return f;
 };
 
 const listFilesLocal = async (config) => {
-  const { directory } = config;
-  const directoryPath = path.join(__dirname, directory, 'staging/pending');
+  const dir = process.env.LOCAL_SOURCE_BUCKET || __dirname;
+  const { folder } = config;
+  const directoryPath = path.join(dir, folder);
+  if (VERBOSE) console.debug(`Fetching local file list from '${directoryPath}'`);
   const list = await readdir(directoryPath);
   const files = list.map( filename => ({
     source: 'local',
@@ -201,8 +243,11 @@ const fetchFile = async file => {
 const fetchFileLocal = async file => {
   const filepath = file.path;
   const data = [];
+  //console.log(file)
+  const stripBomStream = await import('strip-bom-stream');
   return new Promise((resolve, reject) => {
 	  fs.createReadStream(filepath)
+      .pipe(stripBomStream.default())
       .pipe(csv())
       .on('data', row => data.push(row))
       .on('end', () => {
@@ -211,13 +256,15 @@ const fetchFileLocal = async file => {
   });
 };
 
-const fetchFileGoogleBucket = file => {
+const fetchFileGoogleBucket = async file => {
   const data = [];
+  const stripBomStream = await import('strip-bom-stream');
   return new Promise((resolve, reject) => {
     storage
 	    .bucket(file.bucket)
 	    .file(file.path)
 	    .createReadStream() //stream is created
+      .pipe(stripBomStream.default())
       .pipe(csv())
       .on('data', row => data.push(row))
       .on('end', () => {
@@ -235,8 +282,9 @@ const fetchFileGoogleBucket = file => {
 
 const moveFile = async (file, destDirectory) => {
   const source = file.source;
-  console.debug(`Moving file to ${destDirectory}`, file);
-  if(source == 'google-bucket') {
+  if (DRYRUN) {
+    return file;
+  } else if(source == 'google-bucket') {
 	  return await moveFileGoogleBucket(file, destDirectory);
   } else {
 	  return await moveFileLocal(file, destDirectory);
@@ -244,7 +292,7 @@ const moveFile = async (file, destDirectory) => {
 };
 
 const moveFileGoogleBucket = async (file, destDirectory) => {
-  console.debug(`Moving google file to ${destDirectory}`, file);
+  if (VERBOSE) console.debug(`Moving google file to ${destDirectory}`, file);
   const dest = path.join(destDirectory, file.name);
   await storage.bucket(file.bucket).file(file.path).move(dest);
   file.path = dest;
@@ -252,20 +300,46 @@ const moveFileGoogleBucket = async (file, destDirectory) => {
 };
 
 const moveFileLocal = async (file, destDirectory) => {
+  const orig = file.path;
+  destDirectory = path.join(path.dirname(path.dirname(file.path)), destDirectory);
   const dest = path.join(destDirectory, file.name);
+  if (VERBOSE) console.debug(`Moving local file to ${dest}`);
+  await mv(orig, dest);
   file.path = dest;
   return file;
 };
 
 
+const writeError = async (file) => {
+  const source = file.source;
+  // rename the file for the error
+  const parsed = path.parse(file.path);
+  const dir = path.dirname(parsed.dir);
+  file.path = `${dir}/errors/${parsed.name}_error.txt`;
 
-const getFile = async (Key) => {
-  const Bucket = process.env.BUCKET;
+  if (DRYRUN) {
+    return file;
+  } else if(source == 'google-bucket') {
+	  return await writeErrorGoogleBucket(file);
+  } else {
+	  return await writeErrorLocal(file);
+  }
 };
 
-const putFile = async (data, filepath) => {
-  const Bucket = process.env.BUCKET;
+const writeErrorGoogleBucket = async (file) => {
+  if (VERBOSE) console.debug(`Writing google error to ${file.path}`, file);
+  //const dest = path.join(destDirectory, file.name);
+  const path = file.path.replace('./','');
+  await storage.bucket(file.bucket).file(path).save(file.error);
+  return file;
 };
+
+const writeErrorLocal = async (file) => {
+  if (VERBOSE) console.debug(`Writing local error to ${file.path}`);
+  fs.writeFileSync(file.path, file.error);
+  return file;
+};
+
 
 /**
  * Transform phrase to camel case.
@@ -285,12 +359,41 @@ function toCamelCase(phrase) {
         .join('');
 }
 
+
+/**
+ * Print out JSON station object
+ * @param {obj} station
+ */
+function prettyPrintStation(station) {
+    if (typeof(station) === 'string') {
+        station = JSON.parse(station);
+    }
+    for (const [key, value] of Object.entries(station)) {
+        if (key !== 'sensor_systems') {
+            console.log(`${key}: ${value}`);
+        } else {
+            console.log('Sensor systems');
+            value.map( (ss) => {
+                for (const [ky, vl] of Object.entries(ss)) {
+                    if (ky !== 'sensors') {
+                        console.log(`-- ${ky}: ${vl}`);
+                    } else {
+                        vl.map((s) => console.debug(`---- ${s.sensor_id} - ${s.measurand_parameter} ${s.measurand_unit}`));
+                    }
+                }
+            });
+        }
+    }
+}
+
+
 module.exports = {
   fetchSecret,
   applyCredentials,
   revokeCredentials,
   request,
   VERBOSE,
+  DRYRUN,
   putObject,
   getObject,
   moveFile,
@@ -298,7 +401,9 @@ module.exports = {
   listFiles,
   readJson,
   writeJson,
+  writeError,
   toCamelCase,
   gzip,
   unzip,
+  prettyPrintStation,
 };
